Time: 5 ms (72.03%), Space: 14.3 MB (19.4%) - LeetHub

+++ 0072-edit-distance/0072-edit-distance.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    long long solve(string& word1, string& word2, int ind1, int ind2, vector<vector<long long>> &dp) 
+    int solve(string& word1, string& word2, int ind1, int ind2, vector<vector<int>> &dp) 
     {
         if (ind1 == word1.size())
              return word2.size() - ind2;
@@ -13,9 +13,9 @@ public:
         if (word1[ind1] == word2[ind2]) {
             return dp[ind1][ind2] = solve(word1, word2, ind1 + 1, ind2 + 1, dp);
         } else {
-            long long insertOp = 1 + solve(word1, word2, ind1, ind2 + 1, dp);     // simulate insert
-            long long deleteOp = 1 + solve(word1, word2, ind1 + 1, ind2, dp);     // simulate delete
-            long long replaceOp = 1 + solve(word1, word2, ind1 + 1, ind2 + 1, dp);// simulate replace
+            int insertOp = 1 + solve(word1, word2, ind1, ind2 + 1, dp);     // simulate insert
+            int deleteOp = 1 + solve(word1, word2, ind1 + 1, ind2, dp);     // simulate delete
+            int replaceOp = 1 + solve(word1, word2, ind1 + 1, ind2 + 1, dp);// simulate replace
 
             return dp[ind1][ind2] = min({insertOp, deleteOp, replaceOp});
         }
@@ -23,7 +23,7 @@ public:
 
     int minDistance(string word1, string word2) {
         int n = word1.size(), m = word2.size();
-        vector<vector<long long>> dp(n + 1, vector<long long>(m + 1, -1));
-        return (int)solve(word1, word2, 0, 0, dp);
+        vector<vector<int>> dp(n + 1, vector<int>(m + 1, -1));
+        return solve(word1, word2, 0, 0, dp);
     }
 };
